Script for creating placeholders

+++ hikka/services/files.py
@@ -16,8 +16,9 @@ class FileService:
 
     @classmethod
     def destroy(cls, file: File):
-        if file.uploaded and file.path:
-            fs = storage.init_fs()
-            fs.rm(file.storage())
+        if not file.is_link():
+            if file.uploaded and file.path:
+                fs = storage.init_fs()
+                fs.rm(file.storage())
 
-        file.delete()
+            file.delete()
